misc bux fixes
git-svn-id: svn://svn.caucho.com/resin/trunk/modules/hessian@3911 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ src/com/caucho/hessian/io/JavaDeserializer.java
@@ -109,7 +109,7 @@ public class JavaDeserializer extends AbstractMapDeserializer {
       if (cost < 0 || cost > (1 << 48))
 	cost = 1 << 48;
 
-      cost += param.length << 48;
+      cost += (long) param.length << 48;
 
       if (cost < bestCost) {
         _constructor = constructors[i];
